requirements.txt
mlflow==1.29.0
paho-mqtt==1.6.1
pyyaml==6.0
rdflib==6.2.0
requests==2.28.1
service==0.6.0
numpy==1.23.4
tensorflow==2.10.0
protobuf==3.20.*
